+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@ant-design/charts": "^1.4.2",
     "@babel/core": "^7.19.3",
     "@babel/preset-env": "^7.19.4",
     "@emotion/react": "^11.10.4",
